Abstract away createNftSellOffer
