+++ src/XrplSandbox/NFTokens/nftClient.ts
@@ -2,11 +2,61 @@ import {
   NFTokenAcceptOffer,
   NFTokenCancelOffer,
   NFTokenCreateOffer,
+  NFTokenCreateOfferFlags,
   TxResponse,
   xrpToDrops,
 } from 'xrpl';
 import { Amount } from 'xrpl/dist/npm/models/common';
+import { MS_IN_S, RIPPLE_EPOCH_IN_MS } from '../constants';
 import { StateRefProvider } from '../types';
+
+/**
+ * Note that this transaction type is used for both BUY and SELLs.
+ * The differentiators are the various properties of the payload.
+ *
+ * SELL requires:
+ * - Flag.tfSellToken to be set
+ *
+ * {@link https://xrpl.org/nftokencreateoffer.html}
+ */
+export const createNftSellOffer = async (
+  stateRefProvider: StateRefProvider,
+  {
+    amount,
+    tokenId,
+    destination,
+    expirationISOString,
+  }: {
+    amount: Amount | number;
+    tokenId: string;
+    destination?: string;
+    expirationISOString?: string;
+  }
+): Promise<TxResponse> => {
+  const { client, wallet } = await stateRefProvider();
+  const nfTokenCreateSellOfferPayload: NFTokenCreateOffer = {
+    TransactionType: 'NFTokenCreateOffer',
+    Account: wallet.address,
+    Amount: typeof amount === 'number' ? xrpToDrops(amount) : amount,
+    TokenID: tokenId,
+    Flags: NFTokenCreateOfferFlags.tfSellToken,
+  };
+
+  if (destination) {
+    nfTokenCreateSellOfferPayload.Destination = destination;
+  }
+  if (expirationISOString) {
+    const dateTimeInMs = new Date(expirationISOString).getTime();
+    const differenceInMs = dateTimeInMs - RIPPLE_EPOCH_IN_MS;
+    nfTokenCreateSellOfferPayload.Expiration = Math.floor(
+      differenceInMs / MS_IN_S
+    );
+  }
+
+  return client.submitAndWait(nfTokenCreateSellOfferPayload, {
+    wallet,
+  });
+};
 
 /**
  * Note that this transaction type is used for both BUY and SELLs.
